refactor(frontend): extract solver column constants to fix react-refresh
